--- packages/client/hmi-client/src/page/workflow-playground/Node.vue
@@ -1,11 +1,0 @@
-<template>
-    <div class="node"></div>
-</template>
-
-<style>
-.node {
-    border: 1px solid white;
-}
-</style>
-
-<script setup lang="ts"></script>
