+++ packages/client/hmi-client/src/temp/workflow-playground/Node.vue
@@ -1,0 +1,11 @@
+<template>
+	<div class="node"></div>
+</template>
+
+<style>
+.node {
+	border: 1px solid white;
+}
+</style>
+
+<script setup lang="ts"></script>
